Add default PR template to suggest users get reviews (#276)
As part of the discussion about reducing friction when making docs
changes, we're proposing removing the requirement for approval before
review.

Users creating PRs will still be **prompted** to get approval for
changes via this template, but it will not be enforced. Vercel CI
passing **will** still be enforced.

* If a user receives approval, the PR is merged as normal.
* If a user doesn't receive approval, the PR will be merged but a ticket
will be opened in the [audit trail
repo](https://github.com/sourcegraph/sec-pr-audit-trail/issues?q=is:issue)
repo which the docs team will be responsible for closing, effectively
performing a post-merge review.