Update cloud docs (#319)
<!-- Explain the changes introduced in your PR -->
These docs were misleading. We can support individual keys per-customer
but it's a premium offering, not the default setup. I think adding that
we may also support it only adds confusion.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.