docs: extend troubleshooting for very large repositories (#329)
For https://github.com/sourcegraph/sourcegraph/issues/62295

This PR updates the documentation with more tips for very large
repositories.

There are difficulties with Code Insights where it may run for a while,
and then tell the user that there were incomplete data points. This
probably came from very large repositories not being able to compute
reasonably fast.

In addition to this documentation update I'm working on giving users
more information about which repositories lead to incomplete datapoints:
https://github.com/sourcegraph/sourcegraph/issues/62578

---

@sourcegraph/search-platform I poked a bit at the search backend when
gathering this info, and would like to get your input if it's accurate,
and if there may be other improvements to make complex queries run
faster on very large repos :)

@mike-r-mclaughlin Could you review if this new info would be helpful
for customers? I'm planning to expose the repositories that caused
incomplete datapoints with
https://github.com/sourcegraph/sourcegraph/issues/62578. Then a customer
can see which repository didn't compute, pick that one, optimize the
query, and then run the big Code Insight again.