Remove semantic versioning references from update page (#343)
<!-- Explain the changes introduced in your PR -->
Sourcegraph releases do not follow semantic versioning. I'm removing a
reference that says we do

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.