Update Cody messaging in docs (#344)
Updates Cody docs for the new Cody messaging outlined in Notion.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.