chore(rel): add new changelog entry marker (#339)
<!-- Explain the changes introduced in your PR -->

As part of [RFC
912](https://docs.google.com/document/d/1-4kCE7fEBvNx_PpGTK37gNtim36DApCvzhBKPp6eTaw/edit#heading=h.wxuck7z3e74)
implementation, we need to add a specific marker in the changelog to
enable the automation to accurately spot where to introduce the new
content.

As this is a comment, it's completely invisible to viewers of the page
when the docs are served.

If somehow by accident this gets removed, it will break the automation. 

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.