Add structural search instructions (#355)
<!-- Explain the changes introduced in your PR -->
This change adds proper guidance on where to add the structualSearch
property to help administrators get structural search to work properly
and bring back the [ ] to the UI
## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.