Emphasis on user impersonation. (#358)
<!-- Explain the changes introduced in your PR -->
This change adds emphasis on allowing user impersonation through
2-legged auth which was something that was missed by
https://github.com/sourcegraph/accounts/issues/8898.
## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.