docs: update release page for 5.4.5099 release (#391)
<!-- Explain the changes introduced in your PR -->

Update docs to reflect new version in table.
Reformat tables for proper alignment
Use `latest` URL so URL doesn't have to be updated in future.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.