remove mentions of to-be-removed separate Cody natural-language search
See https://github.com/sourcegraph/cody/pull/4506.

This capability is NOT being removed, but it's very rarely used (and only by non-enterprise users) separately. It is almost always used as part of Cody chat, which performs a natural-language search and shows the results in every chat in the context row.