Update supported-models.mdx (#419)
Updated docs to reflect that Enterprise supports Gemini models and
clarified the callout for Claude 3, GPT-4o and Gemini

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.