fix(changelog): regenerate entries for 5.4.5099 (#421)
Changelog generator had a bug where it didn't compute the previous
release to start from properly, which led to an incorrect changelog.

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.