Added sitemap (#423)
<!-- Explain the changes introduced in your PR -->
Added a sitemap file. It can be accessed by visiting '/sitemap.xml'.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.