update screenshots and videos for batch changes doc (#451)
<!-- Explain the changes introduced in your PR -->

This updates screenshots and video assets for Batch Changes docs. A lot
of the UI has changed since most of these docs were written, this is an
update to that.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.