Update docs for keyword search GA (#471)
In addition to keyword search-related updates, this PR makes some "low
hanging fruit" refactors to the search docs structure. I feel this is
important to bundle with the keyword search updates so that our
customers can effectively find information on the change and new syntax.

Main changes:
* Always say "keyword search" when listing search capabilities, instead
of "literal search"
* In search examples, switch to `patterntype=keyword` or `regexp`
instead of older types (`lucky`, `literal`, or `standard`). This was a
big change, and I had to adapt some searches to quote the content "..."
* Remove prominent references to structural search, since it's disabled
by default and we're trying to discourage further use. It's still
discoverable through its own page under "Search Types".
* Low-hanging improvements to page structure, as I found the current one
hard to follow:
  * List "Search Query Syntax" near the top
  * Move "Search Examples" under "Search Query Syntax"
  * Rename "Working with Search" to "Advanced Features"

Closes SPLF-107

---------

Co-authored-by: Maedah Batool <me@MaedahBatool.com>
Co-authored-by: Stefan Hengl <stefan.hengl@gmail.com>
Co-authored-by: Nelson Araujo <nelsonjr@users.noreply.github.com>
Co-authored-by: Quinn Slack <sqs@sourcegraph.com>