Update ignore-context.mdx (#479)
Cleared up some confusion around Cody Ignore file not being available
for Enterprise users. There was a discrepancy at the top of the page and
the bottom of the page. Related
[thread](https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1720213942843199)

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.