Fix typo in CodeCommit SSH Connection section (#483)
<!-- Explain the changes introduced in your PR -->
This update fixes a small typo in step 1 of adding CodeCommit repos in a
Docker container.
## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.