CLI: add documentation for Cody CLI (#502)
Fixes
https://linear.app/sourcegraph/issue/CODY-2659/document-cody-cli-on-docssourcegraphcom

Previously, the Cody CLI was documented in a readme file here. As part
of our effort to make the CLI a more polished product, this PR adds new
documentation on how to install and use Cody CLI that will be available
at the URL https://docs.sourcegraph.com/cody/clients/install-cli

<!-- Explain the changes introduced in your PR -->

I tried to find a color that works OK in both light and dark theme.
Dark             |  Light
:-------------------------:|:-------------------------:
![CleanShot 2024-07-17 at 14 58
50@2x](https://github.com/user-attachments/assets/b6dfdb5c-cfa8-436d-bc22-9d40cba9f86a)
| ![CleanShot 2024-07-17 at 14 58
46@2x](https://github.com/user-attachments/assets/a834a5f2-831d-4d9f-8478-e6772af36a20)




## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.