Add entry on calculating user licenses in Sourcegraph (#520)
<!-- Explain the changes introduced in your PR -->
This update introduces an explanation on how license usage is calculated
in Sourcegraph.
## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.