Callout banner for best practice IAM roles (#523)
<!-- Explain the changes introduced in your PR -->

## Pull Request approval
This PR adds important information related to use of AWS IAM Roles as an
alternative to tokens and username and password options. IAM roles are
currently in the development pipeline so we expect this to change in the
near term future,
Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.