Add docs explaining Perforce labels (#519)
<!-- Explain the changes introduced in your PR -->

Just a short explanation on the Perforce page on when labels are
converted to tags.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.