(fix) update docs page with latest release (#544)
<!-- Explain the changes introduced in your PR -->

Update the docs to refer to the latest Sourcegraph release.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.