add rss route for release feed (#551)
[Part 1 of
DINF-196](https://linear.app/sourcegraph/issue/DINF-196/provide-an-alternative-to-rss-feeds-on-releases)

<!-- Explain the changes introduced in your PR -->

## Test Plan

Run a dev build of the project and navigate to
`localhost:3000/changelog.rss`. You'll be able to access the RSS field.
I've added the tags for the most recent changelog entries.

![CleanShot 2024-08-07 at 18 46
53@2x](https://github.com/user-attachments/assets/f8af631b-bfe2-465b-bda2-92ae89ee54f5)

You won't get a UI similar to the screenshot above, however you'll see
XML in the browser. To view it via a UI like mine, you'll need to
install [a chrome
extension](https://chromewebstore.google.com/detail/rss-feed-reader/pnjaodmkngahhkoihejjehlcdlnohgmp)

![CleanShot 2024-08-07 at 18 59
24@2x](https://github.com/user-attachments/assets/a11db4bf-e89d-4fdd-a515-b94682bacb6f)