Add documentation for the newer modelconfig settings (#555)
This is the same content as in
https://github.com/sourcegraph/docs/pull/540, just with some of the
requested changes applied.

---

In the August release, we will be introducing a new and much more
flexible way for a Sourcegraph administrator to configure the LLM models
used by Cody.

Previously the documentation for Cody Enterprise was found on the
`clients/enable-cody-enterprise.mdx` ("Cody > Installation > Cody
Enterprise" in the nav).

This PR moves all of the LLM-specific documentation into a new file,
`clients/model-configuration.mdx` and a new "Model Configuration" nav
element.

This new page then walks through the specifics of the site configuration
options, and includes the previous documentation in a `# Older style
"Completions" Configuration` section.

Let me know what you think.

- Is this sufficiently clear?
- Are there any scenarios that require more elaboration? (We can easily
devote a dozen pages to examples of various permutations of
configuration, but I wanted to err on the side of being concise. WDYT?)
- Is the call out to "you need to file a support ticket to make this
change if you are a Sourcegraph cloud customer" make sense?

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.