fix(docs): Add 'v' to notes URLs for links to specific versions (#560)
<!-- Explain the changes introduced in your PR -->

Links in the release page didn't go to the exact release notes. Add `v`
in the URL fixes this.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.