Update CodeCommit Callout (#566)
<!-- Explain the changes introduced in your PR -->

This update makes a change to this doc to capture the correct status of
our provisions as referenced by
https://linear.app/sourcegraph/issue/SRC-513/support-aws-iam-roles-for-codecommit-with-assumerole
## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.