Update `pings.mdx` to include collection of ip address (#573)
Updating pings docs to include that we collect ip address of the
instance (`remote_ip` in pings)

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.