Fix typo "unliked" to "unlike" (#497)
Fixed typo in the following line: If not set, unliked with GitHub, it
defaults to true, allowing any GitLab user with access to your instance
to sign up.

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.