docs(batches): drop github apps exclusion (#575)
We now support github apps and can drop the warning that excludes them.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.