Update docker compose guide to new syntax (#583)
Changed docker compose syntax to the new style (docker-compose is
deprecated and docker compose should now be used instead)

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.