update changelog pull request links (#591)
<!-- Explain the changes introduced in your PR -->
Closes DINF-175

THis updates all links in the changelog to the correct one.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.