Update troubleshooting.mdx (#598)
Added a section on how to record a CPU profile to diagnose performance
issues with Cody in VSCode

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.