Update billing active user definition to include rejected autocompletion events (#616)
update billing definition to include rejection of autocompletion events

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.