Add callout for perforce repos converted to git (#626)
<!-- Explain the changes introduced in your PR -->

## Pull Request approval
This PR adds a callout to guide customers who may have converted their
Perforce depots to git but might encounter an issue where code insights
are still not available on those depots.

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.