Clean up entrypoint for Deploy and Upgrade (#629)
<!-- Explain the changes introduced in your PR -->

Add some rewording to the main deployment landing page.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.