Update model-configuration.mdx (#649)
Restructured the docs to make it more obvious that Model Configuration
is in EAP

Resolves comments from @loujar in #645 and clarifies legacy vs EAP
status for Model Config

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.