Remove link for "multi-version upgrades" from AMI upgrade docs (#653)
MVU are not supported so link was confusing, and link pointed to the
upgrade notes page anyways.

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.