Clarify what name to use in alloworgsmap (#667)
We use the "name" field from the GitHub API response to match the name
in alloworgsmap. For example, "Team Name" will work, but "team-name"
won't.

    "name": "Team Name",


https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#list-teams-for-the-authenticated-user


https://sourcegraph.sourcegraph.com/github.com/sourcegraph/sourcegraph@07ca65aec65a8ce528846277127b061e98a825ee/-/blob/cmd/frontend/internal/auth/githuboauth/session.go?L286

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.