Add FAQ for using model of preference on Cody CLI (#675)
<!-- Explain the changes introduced in your PR -->
This update introduces an answer to a question where users would like to
find out if they can use a model of choice to chat with Cody on CLI.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.