Updated editorId Property to editorIds open_in_editor.mdx (#679)
The editorId property has been updated to editorIds. The use of the
editorId property will now result in an error, as it is no longer
supported. Updating the configurations to use the new editorIds list
format to avoid issues.


## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.