Fix: Get Version string from the releaseRegistry (#676)
- **Hackathon attempt to get latest version into docs**
- **POC for Versions being updated from the ReleaseRegistry**
- **Remove Versions.tsx, it's no longer required**
- **Update all instances of "v5.6.2535" to be {CURRENT_VERSION}**
- **remove flake.nix whoops**

<!-- Explain the changes introduced in your PR -->

Our version strings are always out of date because we don't have any
automation to update them. This PR resolves that problem by reaching out
to the ReleaseRegistry at build time to get the latest version string.
It then replaces that string in any code block using the PreCodeBlock
component (which was previously broken and is now fixed).
It also adds an exported const `CURRENT_VERSION_STRING` which can be
used outside of Code Blocks and is similarly populated at build time.

The final stage of this PR is automation that kicks off a build and
releases the docs from the release pipeline itself so nobody has to go
in and do that manually.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.

---------

Co-authored-by: Jacob Pleiness <jdpleiness@users.noreply.github.com>