Update model config docs (#690)
<!-- Explain the changes introduced in your PR -->
Ideally, setting `"cody.enabled": true` in the site config should be
sufficient to enable Cody with the default configuration. However, we
discovered a bug where this isn’t currently the case. Without specifying
the `"completions"` or `"modelConfiguration"` fields, the default Cody
setup doesn’t work as expected. This PR updates the code examples on the
model configuration page to clarify that the `"cody.enabled": true`
field must always be accompanied by either the `"completions"` or
`"modelConfiguration"` fields (for more details see
https://github.com/sourcegraph/sourcegraph/pull/839).

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.