Add guidance for Android Studio for JCEF (#697)
<!-- Explain the changes introduced in your PR -->
This update introduces a workaround for web view JCEF for Android Studio
which isn't supported out of the box and causes Cody not to start and
get stuck on the spinning icon.
## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.