Fix typos and add guidance for Windows users (#698)
<!-- Explain the changes introduced in your PR -->
This doc update fixes some typos for a recent change and also adds a
troubleshooting addition for Windows users having issues logging in.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.