Fix errors (#686)
<!-- Explain the changes introduced in your PR -->

Replaced mention of AWS RDS with Google Cloud SQL on the page for how to
install in Google Cloud.
Fixed error

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.