Update description of telemetry events in docs (#694)
I know this is our outdated docs page, but it is still linked from a few
places and I paused when I saw this

<!-- Explain the changes introduced in your PR -->

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.