fix(docs): fix release versions (#714)
<!-- Explain the changes introduced in your PR -->

Release page was showing wrong versions after recent a recent commit.
Reverted the commit and updated the latest version.

## Pull Request approval

Although pull request approval is not enforced for this repository in
order to reduce friction, merging without a review will generate a
ticket for the docs team to review your changes. So if possible, have
your pull request approved before merging.