docs: remove mentions of package code hosts (#724)
We have no plans to work on this, so we are removing mentions of these package code hosts.

Fixes https://linear.app/sourcegraph/issue/SRC-721/remove-package-host-mentions-from-documentation

## Pull Request approval

Although pull request approval is not enforced for this repository in order to reduce friction, merging without a review will generate a ticket for the docs team to review your changes. So if possible, have your pull request approved before merging.