Fix: Version numbering automatically updates again
We only conditionally replace this _if_ there is a version number in the
code block, because this breaks syntax highlighting.

This is kind of a cheap hack, and we should really be replacing the
strings with version numbers further upstream before they get broken
into syntax highlighting chunks, that way we get to keep both working
version numbers, as well as syntax highlighting.