Fix: Version numbering automatically updates again (#884)
We only conditionally replace this _if_ there is a version number in the
code block, because this breaks syntax highlighting.

This is kind of a cheap hack, and we should really be replacing the
strings with version numbers further upstream before they get broken
into syntax highlighting chunks, that way we get to keep both working
version numbers, as well as syntax highlighting.

I tried messing around with contentlayer to replace these strings
further upstream, but I couldn't figure it out within my timeboxed
session. Hence leaving this PR as it is, because it's fairly important
that correct version numbers be displayed and copied by our customers.

## Pull Request approval

You will need to get your PR approved by at least one member of the
Sourcegraph team. For reviews of docs formatting, styles, and component
usage, please tag the docs team via the #docs Slack channel.