chore: remove notebooks from main navigation and outlines (#1062)
Notebooks have been deprecated and are disabled by default. This removes
them from the main navigation and removes a few incoming links to
notebooks on index pages. This does not remove the notebooks pages so
that customers can still search the docs for notebooks pages, just
cannot browse to them.