Added documentation for tracking changesets' actions  (#1102)
<!-- Explain the changes introduced in your PR -->

## Importing / Tracking Changesets 

I performed some testing on the actions available for imported
changesets in Batch Changes. I was able to do the following:

- **Select Action -> Create Comment:** Batch Changes added the same
comment to all of my changesets. On the Bulk Operation tab, the action
indicated it was completed. I confirmed on GitHub that all PRs had the
new comment on them.
- **Select Action -> Merge Changesets:** Batch Changes merged all my
changesets into main. On the Bulk Operation tab, the action indicated it
was completed. I confirmed that these actions were successfully
reflected on GitHub, as all the PRs were merged into main.
- **Select Action -> Close Changesets:** Batch changes closed all my
changesets on GitHub. On the Bulk Operation tab, it said completed. I
confirmed that these actions were successfully reflected on GitHub, as
the PRs were closed.

Updating the docs page on **tracking changesets** to reflect this
behavior.


## Pull Request approval

You will need to get your PR approved by at least one member of the
Sourcegraph team. For reviews of docs formatting, styles, and component
usage, please tag the docs team via the #docs Slack channel.