Inline Role YAML (#1134)
Since sg/sg is no longer public, the files linked to here no longer
exists.

For the link at the bottom, I changed it so it points to the public
snapshot. There are a lot of files in those folders and I'm not sure if
we want to recreate those files somewhere else.