chore(docs): update file name for arch diagram (#1277)
The architecture diagram generated in the monorepo is actually called
`architecture.svg`, not `sg-architecture.svg`, so just changing it here.
I've confirmed the file is in GCP ready to publish.

We originally tried to [switch the name in the
monorepo](https://github.com/sourcegraph/sourcegraph/pull/6738) to match
the docs repo, but it was turning into a more complicated change than it
would be just to update the docs repo :)