Merge several code monitoring sub pages (#1287)
I found it a bit chaotic to move between several pages for a feature as
simple as code monitoring, so I reduced some redundant content and
consolidated it to a length where I feel ok about having it be a single
page only. I think there's more potential for simplification here, but
this feels like a good improvement regardless.

Test plan: Review.