Removed broken links on Telemetry page - added link to admin page (#1355)
Issue:

The last paragraph of the [Telemetry
page](https://sourcegraph.com/docs/admin/telemetry) contains two broken
links that point to legacy docs.

This PR removes the paragraph and replaces it with a sentence directing
users to contact support for more information about telemetry.

Additionally, I noticed that customers could only find the telemetry
from the search bar or the technical changelog. So I added a link to the
telemetry page in the list of admin resources on /docs/admin.

Tested it by running local dev