Manually fix changelog (#1446)
For some reason there was a development release of 6.10.0 that mucked up
the changelog generation (additionally the release registry since 6.10.0
was using `git_sha` `HEAD` rather than a real commit, which messes with
the changelog tool). This is now manually fixed in the registry and
these entries were generated by manually running the changelog gen
binary.