deepsearch/api: remove 'failed' conversation status (#1457)
This status doesn't exist. After poking at it for a while in
https://github.com/sourcegraph/sourcegraph/pull/7964 I've decided the
change is a bit too pervasive (since we store external API shape
directly into DB), and it's probably better to just fix the docs.

Sadly, it looks like there's no published schema for this, so I also
added what the error behaviour is supposed to be.

Part of https://linear.app/sourcegraph/issue/CU-1495/api-housekeeping