Several docs improvements (#1480)
Check out the individual commits as well

- Add deep search MCP config to repo
- Add tooling to set sitemap.xml priority via frontmatter
- Make separate section in sidebar for all tools 

Groups everything that integrates with Sourcegraph into one bucket, and
puts all APIs next to each other.

- Render current version as "Latest" in the version selector
- Make visited links a little easier to read
- Add some more SEO priority frontmatter tags for most important pages
- Deemphasize the experimental Cody CLI
- Add robots.txt file 

Allows to scrape the latest docs and advertises the sitemap.xml, and
disallows to scrape any versioned docs.

- Add dynamically rendered og-images based on document title
- Improve LLM support for docs pages 
    
Adds a Copy as markdown button so it can be pasted into an LLM, and a
`.md` suffix for all pages that returns the markdown version of it, so
it can be fed to LLMs that can read chunk by chunk.


- Improve rendering of images 

Makes it easier to look at the rather small inline screenshots by adding
a zoom feature.

- Drop unused versioning subpath 

    Not in use, the versioning mechanism has been changed.